+++ pwnlib/data/elf/__init__.py
@@ -2,7 +2,6 @@ from __future__ import absolute_import
 from pwnlib.data.elf import fmtstr
 from pwnlib.data.elf import relro
 from pwnlib.data.elf import ret2dlresolve
-from pwnlib.data.elf import ssh_libs
 
 import os
 path = os.path.dirname(__file__)