+++ pwnlib/data/elf/ssh_libs/__init__.py
@@ -1,6 +1,0 @@
-import os
-path = os.path.dirname(__file__)
-
-def get(x):
-    return os.path.join(path, x)
-